renamed old DeviceNanoAction to DeviceNanoXAction
