+++ src/components/DeviceNanoXAction.js
