+++ src/screens/FirmwareUpdate/02-CheckId.js
@@ -13,7 +13,7 @@ import { deviceNames } from "../../wording";
 import colors from "../../colors";
 import StepHeader from "../../components/StepHeader";
 import LText from "../../components/LText";
-import DeviceNanoAction from "../../components/DeviceNanoAction";
+import DeviceNanoXAction from "../../components/DeviceNanoXAction";
 import LiveLogo from "../../icons/LiveLogoIcon";
 import Spinning from "../../components/Spinning";
 import FirmwareProgress from "../../components/FirmwareProgress";
@@ -103,7 +103,7 @@ class FirmwareUpdateCheckId extends Component<Props, State> {
         <TrackScreen category="FirmwareUpdate" name="CheckId" />
         <View style={styles.body}>
           <View style={styles.device}>
-            <DeviceNanoAction
+            <DeviceNanoXAction
               powerAction
               action
               screen="validation"
