+++ src/screens/Manager/Device.js
@@ -13,7 +13,7 @@ import type {
 } from "@ledgerhq/live-common/lib/types/manager";
 import manager from "@ledgerhq/live-common/lib/manager";
 import { removeKnownDevice } from "../../actions/ble";
-import DeviceNano from "../../components/DeviceNanoAction";
+import DeviceNanoX from "../../components/DeviceNanoXAction";
 import LText from "../../components/LText";
 import Space from "../../components/Space";
 import Circle from "../../components/Circle";
@@ -108,7 +108,7 @@ class ManagerDevice extends Component<Props, { opened: boolean }> {
       <ScrollView style={styles.root} contentContainerStyle={styles.container}>
         <TrackScreen category="Manager" name="Device" />
         <View style={styles.device}>
-          <DeviceNano />
+          <DeviceNanoX />
           <LText secondary semiBold style={styles.deviceName}>
             {deviceNames.nanoX.fullDeviceName}
           </LText>
